Fix the Opc.Ua.Types defects found by the read-only audit
Addresses 98 of the 99 findings the read-only audit of the Opc.Ua.Types
assembly reported for Variant/TypeInfo, the identifier types, the four
codecs, NodeState/Nodes, Schema/Utils/Polyfills and the WoT binding, plus
21 further defects found while reviewing those fixes. Every fix carries a
regression test that fails without it.

Variant and TypeInfo
- TryGetDecimal read the narrow signed scalars through m_union.Int32, so a
  negative SByte/Int16 lost its sign.
- Equality and comparison now convert between numeric built-in types, as
  Part 4 7.7.3's implicit conversion and data precedence rules require, and
  CompareTo reports an unrelated pair as not comparable instead of ordering
  it by raw union bits. The comparison operators reject that sentinel rather
  than reading int.MinValue as an ordinary negative result, which otherwise
  made both a < b and b < a true.
- Comparing a null variant against a typed one gave a different answer
  depending on which side it was written on. It is now decided once and
  symmetrically, and a numeric or boolean zero stays unequal to a null
  variant - letting it compare equal collapsed both into one bucket of every
  Dictionary and HashSet keyed on Variant.
- GetHashCode read Float/Double/Enumeration/DateTime through the wrong union
  field and hashed byte arrays by reference, breaking the Equals contract.
- The bitwise operators, Expand, Collapse and ConvertTo handle Enumeration
  and the narrow integer widths instead of silently producing a null.

Identifiers
- NodeId.CompareTo and ExpandedNodeId.CompareTo are antisymmetric against a
  null and against a foreign type; the null case used to report "less than"
  in both directions, which is not even transitive, so a SortedSet could not
  find a NodeId.Null it had just inserted.
- ExpandedNodeId parses svr= as a UInt32 (Part 6 5.2.2.10) through the server
  mapping table, and reports a malformed index instead of silently turning a
  remote node id into a local one.
- NumericRange rejects an empty dimension, renders multi-dimensional ranges,
  and treats a null range as a no-op in every overload.
- DateTimeUtc forces UTC on every parse path (Part 6 5.3.1.6) and saturates
  its arithmetic in both directions on every target framework.
- DiagnosticInfo keeps the UserPermission bit when shifting the
  returnDiagnostics mask down to the operation level (Part 4 7.32).
- QualifiedName.Format emits a namespace prefix when the name itself contains
  a colon, so the text parses back into the namespace it came from.
- XmlElement hashes structurally, matching the structural Equals.

Encoders
- MaxStringLength counts the bytes of the encoded string (Part 3 5.6.4,
  Part 5 6.3.2) in all four codecs. The XML codecs counted UTF-16 code units,
  so one message context gave two answers and a string a binary peer rejects
  survived an XML round trip. The measurement now lives in one place,
  replacing three separate copies.
- The JSON decoder enforces the array, byte string and string limits it
  never had, including on the base64 and shorthand spellings that bypassed
  the checks the audit asked for.
- A Union selector of 0 means no field and an out of range one is an error
  (Part 6 5.2.8, 5.3.7, H.10); the absent SwitchField no longer indexes -1.
- An encoding limit breach is no longer downgraded into a successful decode
  that keeps the raw over-limit body.
- Nested codecs inherit the namespace and server mappings of the message
  they are part of.
- BinaryEncoder.Close no longer leaves a disposed writer behind for Dispose
  to flush, and EncodeableFactory normalizes a namespace zero URI on lookup
  as well as on registration, without collapsing a server scoped id onto the
  local entry.

NodeState and Nodes
- Node.Write compares the attribute's built-in type and value rank, so an
  array no longer reaches a cast that throws InvalidCastException out of a
  method that reports failure as a status code, and the structured
  permission attributes became writable at all.
- CopyTo carries the optional storage and reports the source's change mask
  rather than the bits its own setters raise.
- MinimumSamplingInterval is written as the Duration it is (Part 3 5.6.2).

Utils, Schema and WoT
- CoreUtils.Match rejects an unmatched trailing character without breaking
  the single character wildcard, and no longer reads past an unterminated
  character set.
- The culture aware string.Replace polyfill advances by the length that
  actually matched; under a culture sensitive comparison a match can span a
  different number of characters than the value searched for.
- The virtual file system keeps the bytes a stream never wrote.
- The WoT resolvers unescape the portable nsu= namespace form the way every
  other reader of the same member does, and the NodeSet converter matches a
  reference type by alias or numeric id consistently.

Notably NOT changed: NodeState.Initialize still does not carry over the
optional storage that CopyTo copies. It instantiates a node from a prototype
rather than cloning it, and propagating the prototype's RolePermissions makes
the GDS certificate group nodes unreadable for users that could read them
before. Whether an instance inherits its prototype's permissions is a
security decision for the address space, so the omission is now documented
instead of quietly closed.

+++ src/Opc.Ua.Types/Nodes/VariableNode.cs
@@ -377,7 +377,8 @@ namespace Opc.Ua
                     AccessLevelEx = (uint)value;
                     return ServiceResult.Good;
                 case Attributes.MinimumSamplingInterval:
-                    MinimumSamplingInterval = (int)value;
+                    // The attribute is a Duration (Double), not an Int32.
+                    MinimumSamplingInterval = (double)value;
                     return ServiceResult.Good;
                 case Attributes.Historizing:
                     Historizing = (bool)value;
